widget action menu

+++ client/components/forms/Remove.jsx
@@ -1,43 +1,15 @@
 import React from 'react';
-import Pref from '/client/global/pref.js';
 import { toast } from 'react-toastify';
 
-import ModelSmall from '../smallUi/ModelSmall';
+import ModelNative from '/client/layouts/Models/ModelNative';
 
 // requires Data
 // action : a string to determine which method to use
 // title : a string of the thing you want to delete
 // check : a string for a conformation
 // entry : the object to be deleted
-
-const RemoveWrapper = ({ action, entry, title, check, lockOut })=> {
-  
-  const auth = Roles.userIsInRole(Meteor.userId(), 'remove');
-  const aT = !auth ? Pref.norole : '';
-  const lT = lockOut ? lockOut : '';
-  const btitle = auth && !lockOut ? `delete "${title}"` : `${aT}\n${lT}`;
-  
-  return(
-    <ModelSmall
-      button='Delete'
-      title={btitle}
-      color='redT'
-      icon='fa-minus-circle'
-      lock={!auth || !check || lockOut}
-    >
-      <Remove
-        action={action}
-        entry={entry}
-        title={title}
-        check={check}
-      />
-    </ModelSmall>
-  );   
-};
-
-export default RemoveWrapper;  
       
-const Remove = ({ action, entry, title, check })=> {
+const Remove = ({ action, entry, title, check, access })=> {
   
   function handleRemove(e) {
     e.preventDefault();
@@ -92,29 +64,42 @@ const Remove = ({ action, entry, title, check })=> {
   }
   
   let checkshort = check.split('T')[0];
-
+  
+  if(!access) {
+    return null;
+  }
+  
   return(
-    <div className='actionBox centre centreText'>
-      <p>Are you sure you want to try to delete "{title}"?</p>
-      <p>This cannot be undone and could cause unexpected consequences.</p>
-      <br />
-      <p>Enter "<i className='noCopy redT'>{checkshort}</i>" to confirm.</p>
-      <br />
-      <form onSubmit={(e)=>handleRemove(e)} className='inlineForm'>
-        <input
-          type='text'
-          id='confirmInput'
-          placeholder={checkshort}
-          autoFocus={true}
-          className='noCopy'
-          required />
-        <button
-          className='smallAction redHover'
-          type='submit'
-          id='cutGo'
-          disabled={false}>DELETE</button>
-      </form>
-      <br />
-    </div>
+    <ModelNative
+      dialogId={action+'_multi_delete_form'}
+      title={`delete "${title}"`}
+      icon='fa-solid fa-minus-circle'
+      colorT='redT'>
+      
+      <div className='actionBox centre centreText'>
+        <p>Are you sure you want to try to delete "{title}"?</p>
+        <p>This cannot be undone and could cause unexpected consequences.</p>
+        <br />
+        <p>Enter "<i className='noCopy redT'>{checkshort}</i>" to confirm.</p>
+        <br />
+        <form onSubmit={(e)=>handleRemove(e)} className='inlineForm'>
+          <input
+            type='text'
+            id='confirmInput'
+            placeholder={checkshort}
+            autoFocus={true}
+            className='noCopy'
+            required />
+          <button
+            className='action redSolid'
+            type='submit'
+            id='cutGo'
+            disabled={false}>DELETE</button>
+        </form>
+        <br />
+      </div>
+    </ModelNative>
   );
 };
+
+export default Remove;
